Make changes in the unit test codebase 1. Optimise code 2. assemble all code under one directory 	modified:   TestFunction/UnitTest/execute.py 	renamed:    TestFunction/UnitTest/unit_tests/test_chrome.py -> TestFunction/UnitTest/test_chrome.py 	renamed:    TestFunction/UnitTest/unit_tests/test_gecko.py -> TestFunction/UnitTest/test_gecko.py 	renamed:    TestFunction/UnitTest/unit_tests/test_opera.py -> TestFunction/UnitTest/test_opera.py 	deleted:    TestFunction/UnitTest/unit_tests/__init__.py 	renamed:    TestFunction/UnitTest/utility/utility.py -> TestFunction/UnitTest/utility.py 	deleted:    TestFunction/UnitTest/utility/__init__.py
