+++ TestFunction/UnitTest/test_chrome.py
@@ -1,8 +1,9 @@
-from utility.utility import Utility_object
+from utility import Utility_object
 import unittest
 import requests
 import json
 import wget
+import os
 
 class TestChrome(unittest.TestCase):
 
@@ -38,14 +39,22 @@ class TestChrome(unittest.TestCase):
             _releaseVersion = str(response.json())
 
         _url = _data.get('url')
-        print _url
+
         _mac_bin = _data.get('mac_v_64')
 
+        # API url
         _apiURLBuilder = _url+_releaseVersion+"/"+_mac_bin
 
+        # Download file
         _url = requests.get(_apiURLBuilder)
         if _url.status_code == 200:
             wget.download(_apiURLBuilder)
+
+        # Assert for file exists
+        self.assertTrue(os.path.exists('chromedriver_mac64.zip'))
+
+        # Delete file
+        self.utility.delete_file("chromedriver_mac64.zip")
 
     def testChromeDriverLinux(self):
         _data = self.parseJSONChrome()
@@ -59,8 +68,16 @@ class TestChrome(unittest.TestCase):
         _url = _data.get('url')
         _linux_bin = _data.get('linux_v_64')
 
+        # API url
         _apiURLBuilder = _url+_releaseVersion+"/"+_linux_bin
 
+        # Download file
         _url = requests.get(_apiURLBuilder)
         if _url.status_code == 200:
             wget.download(_apiURLBuilder)
+
+        # Assert for file exists
+        self.assertTrue(os.path.exists('chromedriver_linux64.zip'))
+
+        # Delete file
+        self.utility.delete_file("chromedriver_linux64.zip")
